src/backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
SQLAlchemy==2.0.37
APScheduler==3.11.0
aiohttp==3.11.12
asyncpg==0.30.0
fastapi-restful==0.6.0
requests==2.32.3
newspaper3k==0.2.8
lxml_html_clean==0.4.1
python-dotenv==1.0.1